# specs/003-reusable-subagents/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Physical AI & Humanoid Robotics Chatbot API - Subagent Extension
  description: |
    API for the RAG chatbot with subagent support for the Physical AI & Humanoid Robotics book.
    This specification extends the base RAG chatbot API with subagent routing and management endpoints.
  version: 2.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /agents/{agent_name}:
    get:
      tags: [agents]
      summary: Get agent details
      description: Returns detailed information about a specific agent.
      operationId: getAgent
      parameters:
        - name: agent_name
          in: path
          required: true
          schema:
            type: string
          description: Agent identifier (e.g., "glossary", "hardware")
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetailResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/{agent_name}/chat:
    post:
      tags: [agents]
      summary: Chat with a specific agent directly
      description: |
        Sends a query directly to a specific agent, bypassing the router.
        Useful for testing and when the caller knows which agent to use.
      operationId: chatWithAgent
      parameters:
        - name: agent_name
          in: path
          required: true
          schema:
            type: string
          description: Agent identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Agent response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ChatRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
          description: The user's question about the book
          example: "What is a topic in ROS 2?"
        session_id:
          type: string
          description: Optional session ID for multi-turn conversations
          example: "abc123"
        selected_text:
          type: string
          description: Optional text highlighted by the user
        force_agent:
          type: string
          description: Optional - force routing to a specific agent
          example: "glossary"
    ChatResponse:
      type: object
      required:
        - answer
        - query_id
        - session_id
      properties:
        answer:
          type: string
          description: The generated answer based on book content
        citations:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
          description: List of source citations
        query_id:
          type: string
          description: Unique identifier for this query
        session_id:
          type: string
          description: Session ID for follow-up questions
        latency_ms:
          type: integer
          description: Processing time in milliseconds
        agent_used:
          type: string
          description: Name of the agent that handled the query
          example: "glossary"
        routing_confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Confidence score for the routing decision
    Citation:
      type: object
      required:
        - index
        - source
        - title
      properties:
        index:
          type: integer
          minimum: 1
          description: Citation number (1-based)
        source:
          type: string
          description: File path or URL to source document
        title:
          type: string
          description: Document title
        section:
          type: string
          description: Section within the document
        relevance_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: Similarity score (0-1)
...
    AgentDetailResponse:
      type: object
      required:
        - name
        - domain
        - description
        - keywords
        - status
      properties:
        name:
          type: string
        domain:
          type: string
        description:
          type: string
        keywords:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, error]
        stats:
          type: object
          properties:
            queries_handled:
              type: integer
            avg_latency_ms:
              type: integer
            avg_confidence:
              type: number
              format: float
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - INVALID_QUERY
            - NO_CONTENT_FOUND
            - AGENT_NOT_FOUND
            - ROUTING_FAILED
            - SERVICE_UNAVAILABLE
            - INTERNAL_ERROR
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details